+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Sensor/SceneCaptureCamera.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include "Carla/Actor/ActorDefinition.h"
+#include "Carla/Sensor/NetMediaOutput.h"
 #include "Carla/Sensor/PixelReader.h"
 #include "Carla/Sensor/ShaderBasedSensor.h"
 
@@ -20,12 +21,20 @@ class CARLA_API ASceneCaptureCamera : public AShaderBasedSensor
 
 public:
 
+
   static FActorDefinition GetSensorDefinition();
 
   ASceneCaptureCamera(const FObjectInitializer &ObjectInitializer);
 
+  UPROPERTY()
+  UNetMediaOutput* MediaOutput = nullptr;
+
 protected:
 
+
+  void BeginPlay() override;
+  void EndPlay(const EEndPlayReason::Type EndPlayReason) override;
   void PostPhysTick(UWorld *World, ELevelTick TickType, float DeltaSeconds) override;
 
+private:
 };